changed function within int parameter by int64 and provide getReport

+++ api/account_account.go
@@ -12,11 +12,11 @@ func NewAccountAccountService(c *Client) *AccountAccountService {
 	return &AccountAccountService{client: c}
 }
 
-func (svc *AccountAccountService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountAccountService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountAccountModel, name)
 }
 
-func (svc *AccountAccountService) GetByIds(ids []int) (*types.AccountAccounts, error) {
+func (svc *AccountAccountService) GetByIds(ids []int64) (*types.AccountAccounts, error) {
 	a := &types.AccountAccounts{}
 	return a, svc.client.getByIds(types.AccountAccountModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountAccountService) GetAll() (*types.AccountAccounts, error) {
 	return a, svc.client.getAll(types.AccountAccountModel, a)
 }
 
-func (svc *AccountAccountService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountAccountService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountAccountModel, fields, relations)
 }
 
-func (svc *AccountAccountService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountAccountService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountAccountModel, ids, fields, relations)
 }
 
-func (svc *AccountAccountService) Delete(ids []int) error {
+func (svc *AccountAccountService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountAccountModel, ids)
 }
