+++ api/account_account_type.go
@@ -12,11 +12,11 @@ func NewAccountAccountTypeService(c *Client) *AccountAccountTypeService {
 	return &AccountAccountTypeService{client: c}
 }
 
-func (svc *AccountAccountTypeService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountAccountTypeService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountAccountTypeModel, name)
 }
 
-func (svc *AccountAccountTypeService) GetByIds(ids []int) (*types.AccountAccountTypes, error) {
+func (svc *AccountAccountTypeService) GetByIds(ids []int64) (*types.AccountAccountTypes, error) {
 	a := &types.AccountAccountTypes{}
 	return a, svc.client.getByIds(types.AccountAccountTypeModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountAccountTypeService) GetAll() (*types.AccountAccountTypes, erro
 	return a, svc.client.getAll(types.AccountAccountTypeModel, a)
 }
 
-func (svc *AccountAccountTypeService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountAccountTypeService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountAccountTypeModel, fields, relations)
 }
 
-func (svc *AccountAccountTypeService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountAccountTypeService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountAccountTypeModel, ids, fields, relations)
 }
 
-func (svc *AccountAccountTypeService) Delete(ids []int) error {
+func (svc *AccountAccountTypeService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountAccountTypeModel, ids)
 }
